made 3 subapps
