+++ user_api/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from user_api import models
+
+# Register your models here.
+admin.site.register(models.User)
