+++ user_api/urls.py
@@ -1,0 +1,8 @@
+from django.conf.urls import url
+from django.contrib import admin
+from django.urls import path,include
+from .views import *
+
+urlpatterns = [
+    
+]
